requirement.txt
langchain
langchain-community
python-dotenv
fasttext-wheel
streamlit
deep-translator
openai
transformers
torch
